# FirmaMobileMauiApp/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FirmaWebApi",
    "version": "1.0"
  },
  "paths": {
...
    "/api/RodzajTransportu": {
      "get": {
        "tags": [
          "RodzajTransportu"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RodzajTransportu"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RodzajTransportu"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RodzajTransportu"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "RodzajTransportu"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RodzajTransportu"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RodzajTransportu"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RodzajTransportu"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RodzajTransportu"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RodzajTransportu"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RodzajTransportu"
                }
              }
            }
          }
        }
      }
    },
    "/api/RodzajTransportu/{id}": {
      "get": {
        "tags": [
          "RodzajTransportu"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RodzajTransportu"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RodzajTransportu"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RodzajTransportu"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "RodzajTransportu"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RodzajTransportu"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RodzajTransportu"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RodzajTransportu"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "RodzajTransportu"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/SkladnikZK": {
      "get": {
        "tags": [
          "SkladnikZK"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SkladnikZKForView"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SkladnikZKForView"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SkladnikZKForView"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "SkladnikZK"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkladnikZK"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SkladnikZK"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SkladnikZK"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SkladnikZKForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkladnikZKForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkladnikZKForView"
                }
              }
            }
          }
        }
      }
    },
    "/api/SkladnikZK/{id}": {
      "get": {
        "tags": [
          "SkladnikZK"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SkladnikZKForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkladnikZKForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkladnikZKForView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "SkladnikZK"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkladnikZK"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SkladnikZK"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SkladnikZK"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "SkladnikZK"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/SposobPlatnosci": {
      "get": {
        "tags": [
          "SposobPlatnosci"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SposobPlatnosci"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SposobPlatnosci"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SposobPlatnosci"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "SposobPlatnosci"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SposobPlatnosci"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SposobPlatnosci"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SposobPlatnosci"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SposobPlatnosci"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SposobPlatnosci"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SposobPlatnosci"
                }
              }
            }
          }
        }
      }
    },
    "/api/SposobPlatnosci/{id}": {
      "get": {
        "tags": [
          "SposobPlatnosci"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SposobPlatnosci"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SposobPlatnosci"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SposobPlatnosci"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "SposobPlatnosci"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SposobPlatnosci"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SposobPlatnosci"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SposobPlatnosci"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "SposobPlatnosci"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Produkt": {
        "required": [
          "jednostkaMiary",
          "kod",
          "nazwa",
          "producent",
          "stawkaVatSprzedazy",
          "stawkaVatZakupu"
        ],
        "type": "object",
        "properties": {
          "idProduktu": {
            "type": "integer",
            "format": "int32"
          },
          "kod": {
            "minLength": 1,
            "type": "string"
          },
          "nazwa": {
            "minLength": 1,
            "type": "string"
          },
          "dodatkowaNazwa": {
            "type": "string",
            "nullable": true
          },
          "jednostkaMiary": {
            "minLength": 1,
            "type": "string"
          },
          "symbol": {
            "type": "string",
            "nullable": true
          },
          "swW_PKWiU": {
            "type": "string",
            "nullable": true
          },
          "producent": {
            "minLength": 1,
            "type": "string"
          },
          "idTypu": {
            "type": "integer",
            "format": "int32"
          },
          "fotoURL": {
            "type": "string",
            "nullable": true
          },
          "stawkaVatZakupu": {
            "type": "number",
            "format": "double"
          },
          "stawkaVatSprzedazy": {
            "type": "number",
            "format": "double"
          },
          "odwrotneObciazenie": {
            "type": "boolean"
          },
          "podzielonaPlatnosc": {
            "type": "boolean"
          },
          "stawkaCla": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "stwkaAkcyzy": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "kodPcn": {
            "type": "string",
            "nullable": true
          },
          "kraj": {
            "type": "string",
            "nullable": true
          },
          "czasKompletacji": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "czyAktywna": {
            "type": "boolean"
          },
          "idDostawcy": {
            "type": "integer",
            "format": "int32"
          },
          "typProduktu": {
            "$ref": "#/components/schemas/TypProduktu"
          },
          "dostawca": {
            "$ref": "#/components/schemas/Kontrahent"
          },
          "pozycjaZZ": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PozycjaZZ"
            },
            "nullable": true
          },
          "cena": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cena"
            },
            "nullable": true
          },
          "zlecenieKompletacji": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ZlecenieKompletacji"
            },
            "nullable": true
          },
          "produktUbocznyZK": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProduktUbocznyZK"
            },
            "nullable": true
          },
          "skladnikZK": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkladnikZK"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RodzajTransportu": {
        "required": [
          "nazwa"
        ],
        "type": "object",
        "properties": {
          "idRodzajuTransportu": {
            "type": "integer",
            "format": "int32"
          },
          "nazwa": {
            "minLength": 1,
            "type": "string"
          },
          "uwagi": {
            "type": "string",
            "nullable": true
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "zlecenieZakupu": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ZlecenieZakupu"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SkladnikZK": {
        "required": [
          "ilosc"
        ],
        "type": "object",
        "properties": {
          "idSkladnika": {
            "type": "integer",
            "format": "int32"
          },
          "idZleceniaKompletacji": {
            "type": "integer",
            "format": "int32"
          },
          "idProduktu": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ilosc": {
            "type": "number",
            "format": "double"
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "zlecenieKompletacji": {
            "$ref": "#/components/schemas/ZlecenieKompletacji"
          },
          "produkt": {
            "$ref": "#/components/schemas/Produkt"
          }
        },
        "additionalProperties": false
      },
      "SkladnikZKForView": {
        "type": "object",
        "properties": {
          "idSkladnika": {
            "type": "integer",
            "format": "int32"
          },
          "idZleceniaKompletacji": {
            "type": "integer",
            "format": "int32"
          },
          "zlecenieKompletacjiDane": {
            "type": "string",
            "nullable": true
          },
          "idProduktu": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "produktDane": {
            "type": "string",
            "nullable": true
          },
          "ilosc": {
            "type": "number",
            "format": "double"
          },
          "czyAktywny": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "SposobDostawy": {
        "required": [
          "nazwa"
        ],
        "type": "object",
        "properties": {
          "idSposobuDostawy": {
            "type": "integer",
            "format": "int32"
          },
          "nazwa": {
            "minLength": 1,
            "type": "string"
          },
          "uwagi": {
            "type": "string",
            "nullable": true
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "zlecenieZakupu": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ZlecenieZakupu"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SposobPlatnosci": {
        "required": [
          "nazwa"
        ],
        "type": "object",
        "properties": {
          "idSposobuPlatnosci": {
            "type": "integer",
            "format": "int32"
          },
          "nazwa": {
            "minLength": 1,
            "type": "string"
          },
          "uwagi": {
            "type": "string",
            "nullable": true
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "kontrahent": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Kontrahent"
            },
            "nullable": true
          },
          "zlecenieZakupu": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ZlecenieZakupu"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ZlecenieKompletacji": {
        "required": [
          "dataWystawienia",
          "ilosc",
          "oczekiwanaDataRealizacji"
        ],
        "type": "object",
        "properties": {
          "idZleceniaKompletacji": {
            "type": "integer",
            "format": "int32"
          },
          "idKontrahenta": {
            "type": "integer",
            "format": "int32"
          },
          "dataWystawienia": {
            "type": "string",
            "format": "date-time"
          },
          "oczekiwanaDataRealizacji": {
            "type": "string",
            "format": "date-time"
          },
          "potwierdzonaDataRealizacji": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "idProduktu": {
            "type": "integer",
            "format": "int32"
          },
          "ilosc": {
            "type": "number",
            "format": "double"
          },
          "idMontera": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "priorytet": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "czyAktywne": {
            "type": "boolean"
          },
          "czasZlecenia": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "kontrahent": {
            "$ref": "#/components/schemas/Kontrahent"
          },
          "monter": {
            "$ref": "#/components/schemas/Monter"
          },
          "produkt": {
            "$ref": "#/components/schemas/Produkt"
          },
          "produktUbocznyZK": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProduktUbocznyZK"
            },
            "nullable": true
          },
          "skladnikZK": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkladnikZK"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ZlecenieZakupu": {
        "required": [
          "dataWystawienia"
        ],
        "type": "object",
        "properties": {
          "idZleceniaZakupu": {
            "type": "integer",
            "format": "int32"
          },
          "idKontrahenta": {
            "type": "integer",
            "format": "int32"
          },
          "idTransakcji": {
            "type": "integer",
            "format": "int32"
          },
          "idSposobuDostawy": {
            "type": "integer",
            "format": "int32"
          },
          "idRodzajuTransportu": {
            "type": "integer",
            "format": "int32"
          },
          "idSposobuPlatnosci": {
            "type": "integer",
            "format": "int32"
          },
          "nrOferty": {
            "type": "string",
            "nullable": true
          },
          "czyPotwierdzone": {
            "type": "boolean"
          },
          "nrPotwierdzenia": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "totalOrder": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "dataWystawienia": {
            "type": "string",
            "format": "date-time"
          },
          "czyAktywne": {
            "type": "boolean"
          },
          "kontrahent": {
            "$ref": "#/components/schemas/Kontrahent"
          },
          "sposobDostawy": {
            "$ref": "#/components/schemas/SposobDostawy"
          },
          "typTransakcji": {
            "$ref": "#/components/schemas/TypTransakcji"
          },
          "rodzajTransportu": {
            "$ref": "#/components/schemas/RodzajTransportu"
          },
          "sposobPlatnosci": {
            "$ref": "#/components/schemas/SposobPlatnosci"
          },
          "pozycjaZZ": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PozycjaZZ"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}